+++ metadata.rb
@@ -4,15 +4,19 @@ maintainer_email 'dev@escapestudios.com'
 license 'MIT'
 description 'Installs/Configures New Relic'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '2.3.0'
+version '2.19.0'
 
-%w( debian ubuntu redhat centos fedora scientific amazon windows smartos ).each do |os|
+%w( debian ubuntu redhat centos fedora scientific amazon windows smartos oracle ).each do |os|
   supports os
 end
 
-depends 'python'
+source_url 'https://github.com/escapestudios-cookbooks/newrelic' if respond_to?(:source_url)
+issues_url 'https://github.com/escapestudios-cookbooks/newrelic/issues' if respond_to?(:issues_url)
 
-recommends 'curl'
+depends 'python'
+depends 'curl'
+depends 'apt', '~> 2.0'
+depends 'yum', '~> 3.0'
 
 recipe 'newrelic', 'Adds the New Relic repository, installs & configures the New Relic server monitor agent.'
 recipe 'newrelic::repository', 'Adds the New Relic repository.'