+++ recipes/nodejs_agent.rb
@@ -2,27 +2,16 @@
 # Cookbook Name:: newrelic
 # Recipe:: nodejs_agent
 #
-# Copyright 2012-2014, Escape Studios
+# Copyright 2012-2015, Escape Studios
 #
 
-include_recipe 'newrelic::repository'
-
-license = node['newrelic']['application_monitoring']['license']
-
-# install the newrelic.js file into each projects
 node['newrelic']['nodejs_agent']['apps'].each do |nodeapp|
-  execute 'npm-install-nodejs_agent' do
-    cwd nodeapp['app_path']
-    command "npm #{node['newrelic']['nodejs_agent']['agent_action']} newrelic"
-  end
-
-  template "#{nodeapp['app_path']}/newrelic.js" do
-    cookbook node['newrelic']['nodejs_agent']['template']['cookbook']
-    source node['newrelic']['nodejs_agent']['template']['source']
-    variables(
-      :license => license,
-      :app_name => nodeapp['app_name'],
-      :app_log_level => 'INFO'
-    )
+  newrelic_agent_nodejs nodeapp['app_path'] do
+    license NewRelic.application_monitoring_license(node)
+    app_name nodeapp['app_name'] unless nodeapp['app_name'].nil?
+    app_log_level nodeapp['app_log_level'] unless nodeapp['app_log_level'].nil?
+    app_log_filepath nodeapp['app_log_filepath'] unless nodeapp['app_log_filepath'].nil?
+    capture_params nodeapp['capture_params'] unless nodeapp['capture_params'].nil?
+    ignored_params nodeapp['ignored_params'] unless nodeapp['ignored_params'].nil?
   end
 end